// src/app/api/model/payment-contact-dto.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Contact information for payment coordination
 */
export interface PaymentContactDto { 
    /**
     * Email address for payment coordination (always visible)
     */
    email?: string;
    /**
     * Full name of the contact
     */
    name?: string;
    /**
     * Phone number for payment coordination (only if user has enabled sharing)
     */
    phone?: string;
    /**
     * Profile picture URL
     */
    profilePicture?: string;
}